Add commands for user and admin Add hash code md5 method Refactoring code

+++ src/main/java/ru/stepanenko/tm/Command/ProjectRemoveCommand.java
@@ -1,15 +1,19 @@
 package ru.stepanenko.tm.Command;
 
 import ru.stepanenko.tm.api.service.IProjectService;
+import ru.stepanenko.tm.api.service.IUserService;
 import ru.stepanenko.tm.entity.Project;
+import ru.stepanenko.tm.entity.User;
 
 import java.util.Scanner;
 
 public class ProjectRemoveCommand extends AbstractCommand {
     IProjectService projectService;
+    IUserService userService;
 
-    public ProjectRemoveCommand(IProjectService projectService) {
+    public ProjectRemoveCommand(IProjectService projectService, IUserService userService) {
         this.projectService = projectService;
+        this.userService = userService;
     }
 
     @Override
@@ -24,9 +28,18 @@ public class ProjectRemoveCommand extends AbstractCommand {
 
     @Override
     public void execute() {
+        User currentUser = userService.getCurrentUser();
+        if (currentUser == null) {
+            System.out.println("This command available only login user!");
+            return;
+        }
+        if (projectService.findAll(currentUser.getId()).isEmpty()){
+            System.out.println("List of projects is empty!");
+            return;
+        }
         Scanner scanner = new Scanner(System.in);
         System.out.println("List of projects:");
-        projectService.findAll().forEach(e -> System.out.println("id: " + e.getId()));
+        projectService.findAll(currentUser.getId()).forEach(e -> System.out.println("id: " + e.getId()));
         System.out.print("Please input project ID for remove: ");
         String projectID = scanner.nextLine();
         Project project = projectService.remove(projectID);
